backend/requirements.txt
flask
tensorflow
Pillow
flask_cors
transformers
torch torchvision
numpy